Add support for new gl::Texture buffers
This ports to mir's new gl::Texture but keeps the existing
gl::Texturesource implementation as fallback.

I opted to use parent classes with overrides to implement the slightly
different codepaths for each buffer type. Since we don't know the buffer
type on contruction of MirBufferSGTexture it will act as a proxy once it
know the buffer type. MirBufferSGTexture will figure out what type of
buffer is used when it recive the first buffer, and then constructs a
matching class. As texture types will never change once they are
constructed, there is no need to do this vodoo detection on new buffers,
saving us lots of cpu cycles.

I didn't want to introduce gl commands into miral::buffer as this will
be moved to miroil soon.