Stop linking to either libGL or libGLES to allow single build for both
We've done away with binding OpenGL API explicitly at all since commit
52ef7bd9 ("Explicitly setting GL-mode breaks GTK app rendering. ...").
So, the only difference left between the GLES version (AKA -android) and
GL (AKA -desktop) version is the build-time linkage to either of the
library. And, turns out, there's only 2 places that calls to libGL{,ES}
directly. One of them can be replaced with QOpenGLFunctions, and the
other one is simple enough we can just eglGetProcAddress() ourself.

And that's exactly what this commit does. It eliminates having to build
2 copies of the libraries for GL and GLES variant of Qt; the same
compiled libraries should be usable with both. The old CMake options is
deprecated, and the linkage against either lib is removed.